API: imshow make rgba the defaut stage when down-sampling
imshow used to default to interpolating in data space.  That
makes sense for up-sampled images, but fails in odd ways for
down-sampled images.

Here we introduce a new default value for *interpolation_stage*
'antialiased', which changes the interpolation stage to 'rgba'
if the data is downsampled or upsampled less than a factor of three.