Disable clipping in Agg resamplers.
I chose to add macro guards directly in the agg source as that seemed
easier than copy-pasting the whole code.

I chose to bump the tolerance on test_rgba_antialias (as that's
also getting bumped by the interpolation_stage antialias PR) and to
test_pngsuite (which is not particularly relevant), and to update the
other baselines.